+++ CHANGELOG.md
@@ -3,11 +3,12 @@ All notable changes to this project will be documented in this file.
 
 ## [0.7.2]
 ### Add
- - reset gate if one sensor sends sensibly more messages (TBD) 
+ - Reset gate if one sensor sends sensibly more messages, configurable via configuration file  
  - added database recover API (TBD)  
  
  ### Changed
   - Fixed issue with periodic reset  
+   - Periodic reset is now self-reset  
  
 
 ## [0.7.1]